pyproject.toml
[project]
name = "toolinject"
version = "0.1.0"
description = "Benchmark harness for evaluating tool-mediated jailbreaks and prompt injections"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ToolInject Contributors"}
]

dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "mistralai>=0.1.0",
    "tavily-python>=0.3.0",
    "docker>=7.0.0",
    "aiofiles>=23.2.0",
    "xxhash>=3.4.0",
    "pandas>=2.1.0",
]

[project.optional-dependencies]
ocr = [
    "easyocr>=1.7.0",
    "pillow>=10.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
toolinject = "toolinject.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/toolinject"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
